ログイン状態かつ自身の出品物である場合にのみ削除可能
items_controllerのdestroyアクションを編集

+++ app/controllers/items_controller.rb
@@ -37,8 +37,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if @item.user == current_user
+      if @item.destroy
+        redirect_to root_path
+      end
     end
   end
